Class rename/refactor
The objective of this change is to make each client class name more descriptive of the specific type of API which it was built to access. Originally, "APISession" was chosen for the REST API v2 client because it was modeled after the Requests' "Session" class and it was assumed that would be the only API for which the library would supply a client.

Not only were classes renamed, but it was discovered that the differences between the Events and Change Events API client classes were trivial. Moreover, the documentation for both of these two live under the "V2 Events API" section in the API reference. Hence, the two client classes have been consolidated into "EventsApiV2Client".

The replacements can be made in descending order of length (using sed, for instance) to avoid editing errors from name overlap:

s/ChangeEventsAPISession/EventsApiV2Client/g
s/EventsAPISession/EventsApiV2Client/g
s/APISession/RestApiV2Client/g