BREAKING: move normalize_params into RestApiV2Client override
The behavior is NOT necessary in the generic client, and will likely lead to surprises later in implementations of other APIs.

The use case this breaks is that if any custom client classes were developed based on pagerduty.ApiClient, they will lose the behavior where any query parameters passed through that are list-valued get "[]" automatically appended to the name of the parameter.